refactor: NotificationTab의 target-local 정리
